Name file info.xml directly
When developing apps for Nextcloud it is more usual that as a developer you can directly clone the app and enable it. The makefile is also very specific.

To improve the developer experience I have renamed that file to "info.xml" directly. At least I was kinda confused why I couldn't enable the app the first time I cloned it :-)