fix: Emit warning instead of info when WOIP request is denied
As discussed in #2685 the message s quite important when tring to fix
WOIP issues. The reasoning is that something fails from the perspective of an admin
and it should emit a warning.
Ofc. one could argue that the deny is expected bahaviour and should therefore emit only
an info. Yet I believe that it'd be beneficial overall to consider admins that
struggle to correctly set this up in that case.

Signed-off-by: moanos <julian-samuel@gebuehr.net>