feat: add personal setting to specify the CA chain for document signing
Document signing needs to store keys as richdocuments settings. This
involves the signing key, certificate and the matching CA chain.

As a first step, add code to the personal settings to be able to set a
CA chain that issues the signing key / certificate.

Setting and getting the setting is possible after this; the setting is
not yet exposed in the WOPI CheckFileInfo response.

<https://github.com/CollaboraOnline/online/issues/9992#issuecomment-2396115427>
has instructions on how to generate self-signed certificates for
document signing for development purposes. Related to #4123

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit 20ca5fd77d0753b58c997ad6a7889153d8b2469e)