feat: expose the documentSigningCa personal setting in the WOPI CheckFileInfo
This setting was already possible to read and write from the personal
settings UI, but was not available towards Collabora Online.

Other private user settings like the Zotero API key are exposed in the
WOPI CheckFileInfo reply.

Do the same here: if the feature is enabled in general and this is not a
public share, then include the signature CA setting in the CheckFileInfo
response.

The same still needs doing for the signature cert/key. Related to #4123

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>