Handle template source for guest users
This fixes #5024 where documents created by guest users (using a public
share link) are reset to the template contents every time they are open.

This is due to the template entry in `richdocuments_template` not being
properly deleted when retrieved for the initial WOPI token. A fix was
merged for the matching select but did not fix the delete statement.

See also: 1833a05611853c4f138a1413618ca51aa4802204

Signed-off-by: kaiyou <dev@kaiyou.fr>