fix(mobile): pass format to SaveAs
With direct access, used in the mobile app, `args.format` was typoed as
`args.Format`, leading to format always being undefined. This meant that
the save as popup would use the extension of the current file rather
than the extension you were trying to save as. You could workaround by
editing this manually, but it was annoying and not working as intended.

The browser version wasn't broken, as it uses a different code path to
save documents (see src/mixins/saveAs.js)

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>